Refactor sub clients directory structure
