+++ src/pypaystack2/paystack.py
@@ -1,6 +1,6 @@
 from pypaystack2.base_api_client import BaseAPIClient, BaseAsyncAPIClient
-from pypaystack2.sub_clients.apple_pay import AsyncApplePayClient, ApplePayClient
-from pypaystack2.sub_clients.bulk_charges import AsyncBulkChargeClient, BulkChargeClient
+from pypaystack2.sub_clients.async_clients.apple_pay import AsyncApplePayClient
+from pypaystack2.sub_clients.async_clients.bulk_charges import AsyncBulkChargeClient
 from pypaystack2.sub_clients.charge import AsyncChargeClient, ChargeClient
 from pypaystack2.sub_clients.customers import AsyncCustomerClient, CustomerClient
 from pypaystack2.sub_clients.dedicated_accounts import (
@@ -36,6 +36,12 @@ from pypaystack2.sub_clients.subaccounts import AsyncSubAccountClient, SubAccoun
 from pypaystack2.sub_clients.subscriptions import (
     AsyncSubscriptionClient,
     SubscriptionClient,
+)
+from pypaystack2.sub_clients.sync_clients.apple_pay import (
+    ApplePayClient,
+)
+from pypaystack2.sub_clients.sync_clients.bulk_charges import (
+    BulkChargeClient,
 )
 from pypaystack2.sub_clients.terminals import TerminalClient, AsyncTerminalClient
 from pypaystack2.sub_clients.transactions import (
